Resolve Admin/User SAC Removal Mismatch (#5856)
* Disallow access to records flagged for removal

* 1. Allow admins to flag any non-accepted record for removal
2. Ensure action on a SAC also flags the associated audit

* Clarity on an error message

* When reverting a SAC to in_progress, also attempt to get the audit

* Mixin fix. Oops.

* Move RemoveSubmissionView tests to their own file.

One of many.

* Handle new access case

* Linting, tests, and the associated light refactoring.
